+++ src/server/devices/espresense/trilaterationPointDistance.ts
@@ -1,0 +1,6 @@
+export class TrilaterationPointDistance {
+  constructor(
+    public readonly pointName: string,
+    public readonly distance: number,
+  ) {}
+}