+++ src/server/devices/espresense/trilaterationRatedCoordinate.ts
@@ -1,0 +1,7 @@
+export class TrilaterationRatedCoordinate {
+  constructor(
+    public coordinateName: string,
+    public rating: number,
+    public matchCount: number = 1,
+  ) {}
+}